nexus-backend/requirements.txt
# ── Core Web Framework ──────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12

# ── Database / Storage ──────────────────────────────────────────────────────
supabase==2.9.0
redis==5.1.0
pgvector==0.3.5            # Semantic memory via Supabase pgvector

# ── AI Providers ────────────────────────────────────────────────────────────
google-generativeai==0.8.3  # Gemini 2.5 Flash
groq==0.11.0
anthropic==0.37.1
openai==1.52.0
mistralai==1.1.0
httpx==0.27.2

# ── Agentic Frameworks ──────────────────────────────────────────────────────
langgraph==0.2.35
crewai==0.80.0
langsmith==0.1.130
agentops==0.3.20            # Traces / evals

# ── Genetic Evolution ───────────────────────────────────────────────────────
deap==1.4.1                 # DEAP genetic algorithms for MARS evolution
numpy==1.26.4

# ── Cryptography (Sprint 6: RSA-2048 improvement proof signing) ─────────────
cryptography>=42.0.0

# ── Observability / Memory ──────────────────────────────────────────────────
mem0ai==0.1.34              # Tiered memory
transformerlens==1.19.0     # Mechanistic interpretability
langchain-core==0.3.10      # LangChain utilities

# ── Scheduling / Background ─────────────────────────────────────────────────
apscheduler==3.10.4
asyncio-mqtt==0.16.2        # Event bus shim

# ── HTTP / Auth ─────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==11.1.0
razorpay==1.4.2             # INR payments (Sprint 2)

# ── Dev / Test ──────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx                       # Already above
black==24.10.0
ruff==0.7.0
